+++ src/layers/dropout.c
@@ -2,15 +2,15 @@
 #include "../../include/utils.h"
 #include <assert.h>
 
-DropoutLayer dropout_layer_init(const double dropout, const bool eval){
+DropoutLayer dropout_layer_init(const float dropout, const bool eval){
     DropoutLayer dropout_layer;
     dropout_layer.dropout = dropout;
     dropout_layer.eval = eval;
     return dropout_layer;
 }
 
-static inline double dropout(double x, double dropout){
-    double u = rand_uniform(0.0, 1.0);
+static inline float dropout(float x, float dropout){
+    float u = rand_uniform(0.0, 1.0);
     if(u < dropout) return 0.0;
     else{
         return x / (1.0 - dropout);
@@ -22,7 +22,7 @@ void dropout_layer_forward(const DropoutLayer *dropout_layer, Tensor *tensor){
         for(size_t i = 0; i < tensor->shape[0]; i++){
             for(size_t j = 0; j < tensor->shape[1]; j++){
                 for(size_t k = 0; k < tensor->shape[2]; k++){
-                    double x = tensor_get_elem(tensor, (uint32_t[]){i, j, k});
+                    float x = tensor_get_elem(tensor, (uint32_t[]){i, j, k});
                     tensor_put_elem(tensor, (uint32_t[]){i, j, k}, dropout(x, dropout_layer->dropout));
                 }   
             }
