+++ test/test_blas.c
@@ -7,23 +7,23 @@ int main() {
     //Taking dot product
 
     int m = 2, k = 3, n = 2;
-    double alpha = 1.0, beta = 0.0;
+    float alpha = 1.0, beta = 0.0;
 
     // dynamically allocate
-    void *A = malloc(m * k * sizeof(double));
-    void *B = malloc(k * n * sizeof(double));
-    double *C = malloc(m * n * sizeof(double));
+    void *A = malloc(m * k * sizeof(float));
+    void *B = malloc(k * n * sizeof(float));
+    float *C = malloc(m * n * sizeof(float));
 
     // initialize A
-    double tmpA[6] = {1, 2, 3,
+    float tmpA[6] = {1, 2, 3,
                       4, 5, 6};
-    for (int i = 0; i < 6; i++) ((double*)A)[i] = tmpA[i];
+    for (int i = 0; i < 6; i++) ((float*)A)[i] = tmpA[i];
 
     // initialize B
-    double tmpB[6] = {7, 8,
+    float tmpB[6] = {7, 8,
                       9, 10,
                       11, 12};
-    for (int i = 0; i < 6; i++) ((double*)B)[i] = tmpB[i];
+    for (int i = 0; i < 6; i++) ((float*)B)[i] = tmpB[i];
 
     // initialize C
     for (int i = 0; i < m*n; i++) C[i] = 0.0;
